tamer_ocr/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0